+++ srcs/parse/parse_lst_function.c
@@ -3,7 +3,6 @@
 int	cmd_lst_add(t_cmd_lst *nil, t_token *token)
 {
 	t_cmd_lst	*new;
-	t_cmd_lst *current;
 
 	new = (t_cmd_lst *)malloc(sizeof(t_cmd_lst));
 	if (!new)
@@ -15,22 +14,13 @@ int	cmd_lst_add(t_cmd_lst *nil, t_token *token)
 	new->prev = nil->prev;
 	new->next = nil;
 	nil->prev = new;
-
-	current = nil->next;
-	while (current != nil)
-	{
-		printf("cmd->str = %s\n", current->str);
-		current = current->next;
-	}
 	return (EXIT_SUCCESS);
 }
 
 int	redirect_lst_add(t_redirect *nil, t_token *tokens)
 {
 	t_redirect	*new;
-	t_redirect	*current;
 
-	printf("------redirect_lst_add_start-----\n");
 	new = (t_redirect *)malloc(sizeof(t_redirect));
 	if (!new)
 		return (EXIT_FAILURE);
@@ -45,18 +35,10 @@ int	redirect_lst_add(t_redirect *nil, t_token *tokens)
 	new->prev = nil->prev;
 	new->next = nil;
 	nil->prev = new;
-	current = nil->next;
-	while (current != nil)
-	{
-		printf("redirect->str = %s\n", current->str);
-		current = current->next;
-	}
-	printf("------redirect_lst_add_end-----\n");
-	printf("\n");
 	return (EXIT_SUCCESS);
 }
 
-int	nlst_add(t_nlst *nil)
+int	nlst_add(t_nlst *nil, t_envlist *env_lst)
 {
 	t_nlst	*new;
 
@@ -65,6 +47,7 @@ int	nlst_add(t_nlst *nil)
 		return (EXIT_FAILURE);
 	new->cmd = init_cmd_lst();
 	new->redirect = init_redirect();
+	new->envp_lst = env_lst;
 	if (!new->cmd || !new->redirect)
 		return (EXIT_FAILURE);
 	nil->prev->next = new;
